// webshop-components/src/components/Table/interfaces.ts
export enum ColumnAlignment {
  Center = 'Center',
  Left = 'Left',
  Right = 'Right',
}

export enum RowState {
  GreenText = 'GreenText',
  RedText = 'RedText',
  Grey = 'Grey',
  Error = 'Error',
  Warning = 'Warning',
  CrossedOut = 'CrossedOut',
  Default = 'Default',
}

export enum CellStatus {
  Warning = 'Warning',
  Danger = 'Danger',
  Success = 'Success',
  Info = 'Info',
}

export interface IStatusDefinition {
  [statusValue: string]: {
    label: string;
    type?: CellStatus;
    color?: string;
  };
}

export interface IGetRowState<T> {
  (row: IRowInfo<T>): RowState;
}

export interface IColumnAccessorFunction<T, R = string> {
  (data: T): R;
}

export interface IColumnAggregateFunction<V> {
  (values: V[]): V;
}

export interface IFilterFunction {
  (query: { pivotId?: string, id?: string, value: string }, row: any, column: any): boolean;
}

export interface IColumnCellFormatFunction {
  <T>(value: T, row: any): string | JSX.Element; // TODO consult @domagoj.cerjan about types here
}
export interface IColumnCellFormatFunction {
  <T>(value: T, row: any): string | JSX.Element; // TODO consult @domagoj.cerjan about types here
}

export interface ICellEventHandler<T> {
  (event: React.SyntheticEvent<HTMLDivElement | HTMLButtonElement>, row: T, original: IRowInfo<T>): void;
}
// ...
export interface IColumnConfig<T> {
  Cell?: any;
  Footer?: ComponentConstructor<any> | string;
  accessor?: DeepKeyOf<T> | IColumnAccessorFunction<T, any>;
  aggregate?: IColumnAggregateFunction<any>; // #TODO @bigd -> can this be typed correctly?
  alignment?: ColumnAlignment;
  cellType?: CellType;
  className?: string;
  dateTimeFormat?: string;
  field?: DeepKeyOf<T> | string;
  filterable?: boolean;
  filterMethod?: IFilterFunction;
  fixedWidth?: boolean; // #TODO @bigd -> implement it down the chain
  footerLabel?: string;
  footerType?: FooterType;
  formatter?: IColumnCellFormatFunction;
  minWidth?: number;
  maxWidth?: number;
  sortable?: boolean;
  style?: any;
  title: string;
  weight?: number;
  isExternal?: boolean | ((row: T) => boolean);
  actions?: Array<IActionInfo<T>>;
  onClick?: ICellEventHandler<T>;
  onChange?: ICellEventHandler<T>;
  onBlur?: ICellEventHandler<T>;
  url?: string;
  openInNewTab?: boolean;
  constructUrl?: IConstructUrl<T>;
  statusDefinition?: IStatusDefinition;
  urlHandle?: boolean; // added for the link cell
}

export type IRowConfig<T> = Array<IColumnConfig<T>>;
// ...
export interface IActionInfo<T> {
  title?: string | ((original: T) => string);
  tooltip?: string | ((original: T) => string);
  icon?: string;
  className?: string;
  isVisible?: boolean | ((row: any) => boolean);
  disabled?: boolean | ((row: any) => boolean);
  onClick?: ICellEventHandler<T>;
}
// ...
export interface IConstructUrl<T> {
  (original: T, row: IRowInfo<T>): string;
}